Version forked for a review
Version forked to be used in a review on the FGM and the data because it has too much details on the historical part and the theoretical part and not enough on the particular appliaction on the project. Especially for an internal note made to be used as a base for dicussion on the protocole of the data.